separate import path for shared mocks
