+++ packages/javascript/bh-shared-ui/package.json
@@ -113,6 +113,7 @@
     },
     "exports": {
         ".": "./dist/index.js",
+        "./mocks": "./dist/mocks.js",
         "./FileIngest": "./dist/FileIngest.js",
         "./ApiExplorer": "./dist/ApiExplorer.js",
         "./SSOConfiguration": "./dist/SSOConfiguration.js",
