+++ packages/javascript/bh-shared-ui/src/mocks.ts
@@ -1,0 +1,1 @@
+export * from './mocks';
